Объединение окон для работы с контейнером
